requirements.txt
discord.py==1.7.3
discord_webhook==0.16.2
requests==2.27.1
